smart_irrigation/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
python-dotenv==1.0.1
boto3==1.34.162
PyJWT==2.9.0
